pyproject.toml
[tool.poetry]
name = "mynorm"
version = "2.0"
description = ""
authors = ["Pavel Piasetski <pavel.piasetski.web@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.2"
psycopg2 = "^2.9.10"
django-allauth = "^65.7.0"
dj-rest-auth = "^7.0.1"
djangorestframework = "^3.16.0"
requests = "^2.32.3"
djangorestframework-simplejwt = "^5.5.0"
django-ordered-model = "^3.7.4"
django-nested-admin = "^4.1.1"
django-admin-sortable2 = "^2.2.6"
pillow = "^11.2.1"
django-filter = "^25.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
